#5 general refactor
