+++ screens/__init__.py
@@ -1,2 +1,6 @@
+"""
+A Place for the UI code
+"""
+
 from .result_table import ResultTable
 from .RawInfoScreen import RawInfoScreen
